Split urls into urls.py and urls_partials.py
