+++ myapp/urls_partials.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from . import views_partials
+
+urlpatterns = [
+    path('cities', views_partials.cities, name='cities'),
+]
