+++ myapp/utils/trains.py
@@ -1,11 +1,11 @@
 import requests
 from bs4 import BeautifulSoup
 
-url = "https://www.interrail.eu/en/plan-your-trip/interrail-timetable?ol=LONDON+KINGS+CROSS+%28UNITED+KINGDOM%29&ov=007061210&dl=NEWCASTLE+%28UNITED+KINGDOM%29&dv=007084870&vl=&vv=&t=1676116800000&ar=false&rt=&tt=&mc=&mct=0#/"
+url = "https://www.rome2rio.com/map/Split/Ancona"
 
 response = requests.get(url)
 soup = BeautifulSoup(response.text, 'html.parser')
+deeplink_meta = soup.find("meta", attrs={"id": "deeplinkTrip"})
 
-journey_cards = soup.find("div", class_="journey__cards")
-print(journey_cards)
-
+with open("output.txt", "w") as file:
+    file.write(str(soup))
